+++ lua/config/plugins/nvim-autopairs.lua
@@ -1,3 +1,8 @@
-require('nvim-autopairs').setup({
-  map_cr = false,
-})
+return {
+  'windwp/nvim-autopairs',
+  config = function()
+    require('nvim-autopairs').setup({
+        map_cr = false,
+      })
+  end,
+}
