--- lua/config/pp/focus.lua
@@ -1,2 +1,0 @@
-require("focus").setup({
-})
