--- lua/config/pp/formatter.lua
@@ -1,16 +1,0 @@
-require('formatter').setup({
-  logging = false,
-  filetype = {
-    javascript = {
-      -- prettierd
-      function()
-        return {
-          exe = "prettierd",
-          args = {vim.api.nvim_buf_get_name(0)},
-          stdin = true
-        }
-      end
-    },
-    -- other formatters ...
-  }
-})
