--- lua/config/pp/lightspeed.lua
@@ -1,1 +1,0 @@
-require'lightspeed'.setup{}
