--- lua/config/pp/mason-lspconfig.lua
@@ -1,3 +1,0 @@
-require("mason-lspconfig").setup({
-  ensure_installed = { "html", "tsserver", "solargraph", "cssls", "dockerls", "jsonls", "yamlls", "vimls", "rust_analyzer" }
-})
