--- lua/config/pp/mason.lua
@@ -1,10 +1,0 @@
-require("mason").setup({
-  ui = {
-      border = {"┏", "━", "┓", "┃", "┛", "━", "┗", "┃"},
-    icons = {
-      package_installed = "✓",
-      package_pending = "➜",
-      package_uninstalled = "✗"
-    }
-  }
-})
