--- lua/config/pp/minimap.lua
@@ -1,4 +1,0 @@
-vim.g.minimap_width = 10
-vim.g.minimap_auto_start = 0
-vim.g.minimap_auto_start_win_enter = 0
-vim.g.minimap_highlight_range = 1
