--- lua/config/pp/neoscroll.lua
@@ -1,1 +1,0 @@
-require('neoscroll').setup()
