--- lua/config/pp/numb.lua
@@ -1,4 +1,0 @@
-require('numb').setup{
-  show_numbers = true, -- Enable 'number' for the window while peeking
-  show_cursorline = true -- Enable 'cursorline' for the window while peeking
-}
