--- lua/config/pp/nvim-bqf.lua
@@ -1,5 +1,0 @@
-require('bqf').setup({
-  preview = {
-    border_chars = {'┃', '┃', '━', '━', '┏', '┓', '┗', '┛', '█'}
-  },
-})
