--- lua/config/pp/nvim-dap-ui.lua
@@ -1,1 +1,0 @@
-require("dapui").setup()
