--- lua/config/pp/nvim-dap-virtual-text.lua
@@ -1,1 +1,0 @@
-require("nvim-dap-virtual-text").setup()
