--- lua/config/pp/nvim-gps.lua
@@ -1,9 +1,0 @@
-require("nvim-gps").setup({
-  icons = {
-    ["class-name"] = ' ',      -- Classes and class-like objects
-    ["function-name"] = ' ',   -- Functions
-    ["method-name"] = ' ',     -- Methods (functions inside class-like objects)
-    ["container-name"] = '⛶ ',  -- Containers (example: lua tables)
-    ["tag-name"] = '炙'         -- Tags (example: html tags)
-  }
-})
