--- lua/config/pp/nvim-neoclip.lua
@@ -1,1 +1,0 @@
-require('neoclip').setup()
