--- lua/config/pp/nvim-notify.lua
@@ -1,2 +1,0 @@
-vim.notify = require("notify")
-require("notify").setup()
