--- lua/config/pp/nvim-ts-autotag.lua
@@ -1,1 +1,0 @@
-require('nvim-ts-autotag').setup()
