--- lua/config/pp/nvim-ts-rainbow.lua
@@ -1,5 +1,0 @@
-require'nvim-treesitter.configs'.setup {
-  rainbow = {
-    enable = true
-  }
-}
