--- lua/config/pp/todo-comments.lua
@@ -1,1 +1,0 @@
-require("todo-comments").setup {}
