--- lua/config/pp/treesitter.lua
@@ -1,17 +1,0 @@
-require'nvim-treesitter.configs'.setup {
-  markid = { 
-    enable = true,
-    colors = false 
-  },
-  ensure_installed = "all", -- one of "all", "maintained" (parsers with maintainers), or a list of languages
-  highlight = {
-    enable = true,              -- false will disable the whole extension
-  },
-  context_commentstring = {
-    enable = true,
-    enable_autocmd = false,
-  },
-  matchup = {
-    enable = true  
-  }
-}
