--- lua/config/pp/which-key.lua
@@ -1,5 +1,0 @@
-require("which-key").setup({
-  window = {
-    border = { "┏", "━", "┓", "┃", "┛","━", "┗", "┃" },
-  },
-})
